Refactor Code + Marco Bug Fixes

+++ src/main/java/com/SpringField/engine/util/Util.java
@@ -57,45 +57,42 @@ public class Util {
     /*
      * Board
      */
-    public static byte[] tilesResource =
-            new byte[] { DESERT, WOOD, WOOD, WOOD, WOOD, BRICK, BRICK, BRICK, SHEEP, SHEEP, SHEEP, SHEEP, HAY, HAY, HAY,
-                    HAY, ROCK, ROCK, ROCK };
+    public static byte[] tilesResource = new byte[] { DESERT, WOOD, WOOD, WOOD, WOOD, BRICK, BRICK, BRICK, SHEEP, SHEEP,
+            SHEEP, SHEEP, HAY, HAY, HAY, HAY, ROCK, ROCK, ROCK };
 
     public static byte[] tilesNumber = new byte[] { 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 8, 8, 9, 9, 10, 10, 11, 11, 12 };
 
-    public static int[][] edgeToVertex =
-            new int[][] { new int[] { 0, 1 }, new int[] { 1, 2 }, new int[] { 2, 3 }, new int[] { 3, 4 },
-                    new int[] { 4, 5 }, new int[] { 5, 6 }, new int[] { 0, 8 }, new int[] { 2, 10 },
-                    new int[] { 4, 12 }, new int[] { 6, 14 }, new int[] { 7, 8 }, new int[] { 8, 9 },
-                    new int[] { 9, 10 }, new int[] { 10, 11 }, new int[] { 11, 12 }, new int[] { 13, 14 },
-                    new int[] { 14, 15 }, new int[] { 7, 17 }, new int[] { 9, 19 }, new int[] { 11, 21 },
-                    new int[] { 13, 23 }, new int[] { 15, 25 }, new int[] { 16, 17 }, new int[] { 17, 18 },
-                    new int[] { 18, 19 }, new int[] { 19, 20 }, new int[] { 20, 21 }, new int[] { 21, 22 },
-                    new int[] { 22, 23 }, new int[] { 23, 24 }, new int[] { 24, 25 }, new int[] { 25, 26 },
-                    new int[] { 16, 27 }, new int[] { 18, 29 }, new int[] { 20, 31 }, new int[] { 22, 33 },
-                    new int[] { 24, 35 }, new int[] { 26, 37 }, new int[] { 27, 28 }, new int[] { 28, 29 },
-                    new int[] { 29, 30 }, new int[] { 30, 31 }, new int[] { 31, 32 }, new int[] { 32, 33 },
-                    new int[] { 33, 34 }, new int[] { 34, 35 }, new int[] { 35, 36 }, new int[] { 36, 37 },
-                    new int[] { 28, 38 }, new int[] { 30, 40 }, new int[] { 32, 42 }, new int[] { 34, 44 },
-                    new int[] { 36, 46 }, new int[] { 38, 39 }, new int[] { 39, 40 }, new int[] { 40, 41 },
-                    new int[] { 41, 42 }, new int[] { 42, 43 }, new int[] { 43, 44 }, new int[] { 44, 45 },
-                    new int[] { 45, 46 }, new int[] { 39, 47 }, new int[] { 41, 49 }, new int[] { 43, 51 },
-                    new int[] { 45, 53 }, new int[] { 47, 48 }, new int[] { 48, 49 }, new int[] { 49, 50 },
-                    new int[] { 50, 51 }, new int[] { 51, 52 }, new int[] { 52, 53 },
+    public static byte[][] edgeToVertex = new byte[][] { new byte[] { 0, 1 }, new byte[] { 1, 2 }, new byte[] { 2, 3 },
+            new byte[] { 3, 4 }, new byte[] { 4, 5 }, new byte[] { 5, 6 }, new byte[] { 0, 8 }, new byte[] { 2, 10 },
+            new byte[] { 4, 12 }, new byte[] { 6, 14 }, new byte[] { 7, 8 }, new byte[] { 8, 9 }, new byte[] { 9, 10 },
+            new byte[] { 10, 11 }, new byte[] { 11, 12 }, new byte[] { 13, 14 }, new byte[] { 14, 15 },
+            new byte[] { 7, 17 }, new byte[] { 9, 19 }, new byte[] { 11, 21 }, new byte[] { 13, 23 },
+            new byte[] { 15, 25 }, new byte[] { 16, 17 }, new byte[] { 17, 18 }, new byte[] { 18, 19 },
+            new byte[] { 19, 20 }, new byte[] { 20, 21 }, new byte[] { 21, 22 }, new byte[] { 22, 23 },
+            new byte[] { 23, 24 }, new byte[] { 24, 25 }, new byte[] { 25, 26 }, new byte[] { 16, 27 },
+            new byte[] { 18, 29 }, new byte[] { 20, 31 }, new byte[] { 22, 33 }, new byte[] { 24, 35 },
+            new byte[] { 26, 37 }, new byte[] { 27, 28 }, new byte[] { 28, 29 }, new byte[] { 29, 30 },
+            new byte[] { 30, 31 }, new byte[] { 31, 32 }, new byte[] { 32, 33 }, new byte[] { 33, 34 },
+            new byte[] { 34, 35 }, new byte[] { 35, 36 }, new byte[] { 36, 37 }, new byte[] { 28, 38 },
+            new byte[] { 30, 40 }, new byte[] { 32, 42 }, new byte[] { 34, 44 }, new byte[] { 36, 46 },
+            new byte[] { 38, 39 }, new byte[] { 39, 40 }, new byte[] { 40, 41 }, new byte[] { 41, 42 },
+            new byte[] { 42, 43 }, new byte[] { 43, 44 }, new byte[] { 44, 45 }, new byte[] { 45, 46 },
+            new byte[] { 39, 47 }, new byte[] { 41, 49 }, new byte[] { 43, 51 }, new byte[] { 45, 53 },
+            new byte[] { 47, 48 }, new byte[] { 48, 49 }, new byte[] { 49, 50 }, new byte[] { 50, 51 },
+            new byte[] { 51, 52 }, new byte[] { 52, 53 },
 
-            };
+    };
 
-    public static int[][] tileToVertex =
-            new int[][] { new int[] { 0, 1, 2, 8, 9, 10 }, new int[] { 2, 3, 4, 10, 11, 12 },
-                    new int[] { 4, 5, 6, 12, 13, 14 }, new int[] { 7, 8, 9, 17, 18, 19 },
-                    new int[] { 9, 10, 11, 19, 20, 21 }, new int[] { 11, 12, 13, 21, 22, 23 },
-                    new int[] { 13, 14, 15, 23, 24, 25 }, new int[] { 16, 17, 18, 27, 28, 29 },
-                    new int[] { 18, 19, 20, 29, 30, 31 }, new int[] { 20, 21, 22, 31, 32, 33 },
-                    new int[] { 22, 23, 24, 33, 34, 35 }, new int[] { 24, 25, 26, 35, 36, 37 },
-                    new int[] { 28, 29, 30, 38, 39, 40 }, new int[] { 30, 31, 32, 40, 41, 42 },
-                    new int[] { 32, 33, 34, 42, 43, 44 }, new int[] { 34, 35, 36, 44, 45, 46 },
-                    new int[] { 39, 40, 41, 47, 48, 49 }, new int[] { 41, 42, 43, 49, 50, 51 },
-                    new int[] { 43, 44, 45, 51, 52, 53 }, };
+    public static byte[][] tileToVertex = new byte[][] { new byte[] { 0, 1, 2, 8, 9, 10 },
+            new byte[] { 2, 3, 4, 10, 11, 12 }, new byte[] { 4, 5, 6, 12, 13, 14 }, new byte[] { 7, 8, 9, 17, 18, 19 },
+            new byte[] { 9, 10, 11, 19, 20, 21 }, new byte[] { 11, 12, 13, 21, 22, 23 },
+            new byte[] { 13, 14, 15, 23, 24, 25 }, new byte[] { 16, 17, 18, 27, 28, 29 },
+            new byte[] { 18, 19, 20, 29, 30, 31 }, new byte[] { 20, 21, 22, 31, 32, 33 },
+            new byte[] { 22, 23, 24, 33, 34, 35 }, new byte[] { 24, 25, 26, 35, 36, 37 },
+            new byte[] { 28, 29, 30, 38, 39, 40 }, new byte[] { 30, 31, 32, 40, 41, 42 },
+            new byte[] { 32, 33, 34, 42, 43, 44 }, new byte[] { 34, 35, 36, 44, 45, 46 },
+            new byte[] { 39, 40, 41, 47, 48, 49 }, new byte[] { 41, 42, 43, 49, 50, 51 },
+            new byte[] { 43, 44, 45, 51, 52, 53 }, };
 
     public static byte DEFAULT_NUM_TILES = (byte) tilesResource.length;
     public static byte DEFAULT_NUM_VERTICES = 54;
@@ -139,7 +136,7 @@ public class Util {
             nodeToEdgeList[i] = new ArrayList<>();
         }
         for (int i = 0; i < edgeToVertex.length; i++) {
-            int[] verticies = edgeToVertex[i];
+            byte[] verticies = edgeToVertex[i];
             for (int j = 0; j < verticies.length; j++) {
                 nodeToEdgeList[verticies[j]].add((byte) i);
             }
